Добавил метод logout

+++ backend/app/Http/Controllers/AuthController.php
@@ -52,7 +52,9 @@ class AuthController extends Controller
 
   public function logout(Request $request)
   {
-    $request->user()->currentAccessToken()->delete();
+    if ($request->user()) {
+      $request->user()->currentAccessToken()->delete();
+    }
 
     return response()->json([
       'is_logout' => true
